UserSchema and changes
