+++ backend/UserSchema.js
@@ -1,0 +1,25 @@
+const mongoose = require("mongoose");
+
+// Schema for Users
+const Users = mongoose.model("Users", {
+  name: { type: String },
+  email: { type: String, unique: true },
+  password: { type: String },
+  cartData: { type: Object },
+  date: { type: Date, default: Date.now() },
+});
+
+// Schema for Products
+const Product = mongoose.model("Product", {
+  id: { type: Number, required: true },
+  name: { type: String, required: true },
+  description: { type: String, required: true },
+  image: { type: String, required: true },
+  category: { type: String, required: true },
+  new_price: { type: Number },
+  old_price: { type: Number },
+  date: { type: Date, default: Date.now },
+  available: { type: Boolean, default: true },
+});
+
+module.exports = { Users, Product };
