--- backend/index.js
@@ -1,162 +1,0 @@
-const express = require("express");
-const app = express();
-const mongoose = require("mongoose");
-const jwt = require("jsonwebtoken");
-const multer = require("multer");
-const path = require("path");
-const cors = require("cors");
-const port = process.env.PORT || 4000;
-
-app.use(express.json());
-app.use(cors());
-
-mongoose.connect("mongodb+srv://harmeet96:U75SYB05AWP1@cluster0.0elap.mongodb.net/e-commerce");
-
-//Image Storage Engine 
-const storage = multer.diskStorage({
-  destination: './upload/images',
-  filename: (req, file, cb) => {
-    return cb(null, `${file.fieldname}_${Date.now()}${path.extname(file.originalname)}`)
-  }
-})
-const upload = multer({ storage: storage })
-app.post("/upload", upload.single('product'), (req, res) => {
-  res.json({
-    success: 1,
-    image_url: `/images/${req.file.filename}`
-  })
-})
-
-// Route for Images folder
-app.use('/images', express.static('upload/images'));
-
-// MiddleWare to fetch user from token
-const fetchuser = async (req, res, next) => {
-  const token = req.header("auth-token");
-  if (!token) {
-    res.status(401).send({ errors: "Please authenticate using a valid token" });
-  }
-  try {
-    const data = jwt.verify(token, "secret_ecom");
-    req.user = data.user;
-    next();
-  } catch (error) {
-    res.status(401).send({ errors: "Please authenticate using a valid token" });
-  }
-};
-
-// Schema for creating user model
-const Users = mongoose.model("Users", {
-  name: { type: String },
-  email: { type: String, unique: true },
-  password: { type: String },
-  cartData: { type: Object },
-  date: { type: Date, default: Date.now() },
-});
-
-// Schema for creating Product
-const Product = mongoose.model("Product", {
-  id: { type: Number, required: true },
-  name: { type: String, required: true },
-  description: { type: String, required: true },
-  image: { type: String, required: true },
-  category: { type: String, required: true },
-  new_price: { type: Number },
-  old_price: { type: Number },
-  date: { type: Date, default: Date.now },
-  available: { type: Boolean, default: true },
-});
-
-// ROOT API Route For Testing
-app.get("/", (req, res) => {
-  res.send("Root");
-});
-
-// Create an endpoint at ip/login for login the user and giving auth-token
-app.post('/login', async (req, res) => {
-  console.log("Login");
-  let success = false;
-  let user = await Users.findOne({ email: req.body.email });
-  if (user) {
-    const passCompare = req.body.password === user.password;
-    if (passCompare) {
-      const data = {
-        user: {
-          id: user.id
-        }
-      }
-      success = true;
-      console.log(user.id);
-      const token = jwt.sign(data, 'secret_ecom');
-      res.json({ success, token });
-    }
-    else {
-      return res.status(400).json({ success: success, errors: "please try with correct email/password" })
-    }
-  }
-  else {
-    return res.status(400).json({ success: success, errors: "please try with correct email/password" })
-  }
-})
-
-//Create an endpoint at ip/auth for regestring the user & sending auth-token
-app.post('/signup', async (req, res) => {
-  console.log("Sign Up");
-  let success = false;
-  let check = await Users.findOne({ email: req.body.email });
-  if (check) {
-    return res.status(400).json({ success: success, errors: "existing user found with this email" });
-  }
-  let cart = {};
-  for (let i = 0; i < 300; i++) {
-    cart[i] = 0;
-  }
-  const user = new Users({
-    name: req.body.username,
-    email: req.body.email,
-    password: req.body.password,
-    cartData: cart,
-  });
-  await user.save();
-  const data = {
-    user: {
-      id: user.id
-    }
-  }
-
-  const token = jwt.sign(data, 'secret_ecom');
-  success = true;
-  res.json({ success, token })
-})
-
-// Create an endpoint for saving the product in cart
-app.post('/addtocart', fetchuser, async (req, res) => {
-  console.log("Add Cart");
-  let userData = await Users.findOne({ _id: req.user.id });
-  userData.cartData[req.body.itemId] += 1;
-  await Users.findOneAndUpdate({ _id: req.user.id }, { cartData: userData.cartData });
-  res.send("Added")
-})
-
-// Create an endpoint for removing the product in cart
-app.post('/removefromcart', fetchuser, async (req, res) => {
-  console.log("Remove Cart");
-  let userData = await Users.findOne({ _id: req.user.id });
-  if (userData.cartData[req.body.itemId] != 0) {
-    userData.cartData[req.body.itemId] -= 1;
-  }
-  await Users.findOneAndUpdate({ _id: req.user.id }, { cartData: userData.cartData });
-  res.send("Removed");
-})
-
-// Create an endpoint for getting cartdata of user
-app.post('/getcart', fetchuser, async (req, res) => {
-  console.log("Get Cart");
-  let userData = await Users.findOne({ _id: req.user.id });
-  res.json(userData.cartData);
-})
-
-app.listen(port, (error) => {
-  if (!error) console.log("Server Running on port " + port);
-  else console.log("Error : ", error);
-});
